Home

GMT.jl Documentation

Index

Examples

Here's the "Hello World"

And the "Hello Round World"

Color images

Perspective view

GMT.jl Documentation

Index

Functions

The GMT.jl types

The Grid type

The Image type

The DATASET type

The CPT type

The Postscript type

Index

GMT.grdcontour

GMT.grdimage

GMT.grdview

GMT.makecpt

GMT.pscoast

GMT.psscale

GMT.psxy

Functions

GMT.psxy — Function.

psxy(cmd0::String="", arg1=[]; caller=[], data=[], fmt="", K=false, O=false,
     first=true, kwargs...)

reads (x,y) pairs from files [or standard input] and generates PostScript code that will plot lines, polygons, or symbols at those locations on a map.

Full option list at psxy

Parameters

A : straight_lines : – Str – By default, geographic line segments are drawn as great circle arcs. To draw them as straight lines, use the -A flag.

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – 'xmin/xmax/ymin/ymax[+r][+uunit]'. Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : – Str – Give a CPT or specify -Ccolor1,color2[,color3,...] to build a linear continuous CPT from those colors automatically. -C

D : offset : – Str – 'dx/dy' Offset the plot symbol or line locations by the given amounts dx/dy.

E : error_bars : – Str – '[x|y|X|Y][+a][+cl|f][+n][+wcap][+ppen]' Draw symmetrical error bars. -E

F : conn : connection : – Str – '[c|n|r][a|f|s|r|refpoint]' Alter the way points are connected -F

G : fill : markerfacecolor : MarkerFaceColor : – Str – Select color or pattern for filling of symbols or polygons. BUT WARN: the alias 'fill' will set the color of polygons OR symbols but not the two together. If your plot has polygons and symbols, use 'fill' for the polygons and 'markerfacecolor' for filling the symbols. Same applyies for W bellow -G

I : intens : – Str or number – Use the supplied intens value (in the [-1 1] range) to modulate the fill color by simulating illumination.

L : closed_polygon : – Str – '[+b|d|D][+xl|r|x0][+yl|r|y0][+ppen]' Force closed polygons. -L

N : no_clip : –- Str or [] – '[c|r]' Do NOT clip symbols that fall outside map border

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

S : symbol : marker : Marker : – Str – '[symbol][size[u]]' Plot symbols (including vectors, pie slices, fronts, decorated or quoted lines). -S Alternatively select a sub-set of symbols using the aliases: marker or Marker and values:

-, x_dash

+, plus

a, *, star

c, circle

d, diamond

g, octagon

h, hexagon

i, v, inverted_tri

n, pentagon

p, ., point

r, rectangle

s, square

t, ^, triangle

x, cross

y, y_dash

W : line_attribs : markeredgecolor : MarkerEdgeColor : – Str – '[pen][attr]' Set pen attributes for lines or the outline of symbols -W WARNING: the pen attributes will set the pen of polygons OR symbols but not the two together. If your plot has polygons and symbols, use W or line_attribs for the polygons and markeredgecolor or MarkerEdgeColor for filling the symbols. Similar to S above.

U : stamp : – Str or Bool or [] – '[[just]/dx/dy/][c|label]' Draw GMT time stamp logo on plot.

V : verbose : – Bool or Str – '[level]' Select verbosity level -V

X : x_offset : – Str – '[a|c|f|r][x-shift[u]]'

Y : y_offset : – Str – '[a|c|f|r][y-shift[u]]' Shift plot origin. -Y

a : aspatial : – Str –

bi : binary_in : – Str –

di : nodata_in : – Str –

e : patern : – Str –

f : colinfo : – Str –

g : gaps : – Str –

h : headers : – Str –

i : input_col : – Str –

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.pscoast — Function.

pscoast(cmd0::String=""; fmt="", clip=[], K=false, O=false, first=true, kwargs...)

Plot continents, shorelines, rivers, and borders on maps. Plots grayshaded, colored, or textured land-masses [or water-masses] on maps and [optionally] draws coastlines, rivers, and political boundaries. A map projection must be supplied.

Full option list at pscoast

Parameters

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – 'xmin/xmax/ymin/ymax[+r][+uunit]'. Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

A : area : – Str or Number – Features with an area smaller than min_area in km^2 or of hierarchical level that is lower than min_level or higher than max_level will not be plotted. -A

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : river_fill : – Str – Set the shade, color, or pattern for lakes and river-lakes. -C

D : res : resolution : – Str – Selects the resolution of the data set to use ((f)ull, (h)igh, (i)ntermediate, (l)ow, and (c)rude). -D

E : ECW : – Str – Tuple(Str, Str); Tuple("code", (pen)), ex: ("PT",(0.5,"red","–")); Tuple((...),(...),...) Select painting or dumping country polygons from the Digital Chart of the World -E

F : box : – Str – Draws a rectangular border around the map scale or rose. -F

G : land : – Str – Select filling or clipping of “dry” areas. -G

I : rivers : – Str – Draw rivers. Specify the type of rivers and [optionally] append pen attributes. -I

L : map_scale : – Str – Dtraw a map scale.

M : dump : – Str – Dumps a single multisegment ASCII output. No plotting occurs. -M

N : borders : – Str – Draw political boundaries. Specify the type of boundary and [optionally] append pen attributes -N

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

S : water : – Str – Select filling or clipping of “wet” areas. -S

Td : rose` : – Str – Draws a map directional rose on the map at the location defined by the reference and anchor points. -Td

Tm : compass : – Str – Draws a map magnetic rose on the map at the location defined by the reference and anchor points. -Tm

U : stamp : – Str or Bool or [] – '[[just]/dx/dy/][c|label]' Draw GMT time stamp logo on plot.

V : verbose : – Bool or Str – '[level]' Select verbosity level -V

W : shore : – Str – Draw shorelines [Default is no shorelines]. Append pen attributes. -W

X : x_offset : – Str – '[a|c|f|r][x-shift[u]]'

Y : y_offset : – Str – '[a|c|f|r][y-shift[u]]' Shift plot origin. -Y

bo : binary_out : – Str –

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.psscale — Function.

psscale(cmd0::String="", arg1=[]; fmt="", K=false, O=false, first=true, kwargs...)

Plots gray scales or color scales on maps.

Full option list at psscale

D : position : – Str – Defines the reference point on the map for the color scale using one of four coordinate systems. -D

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : cmap : – Str – Name of the CPT (for grd_z only). Alternatively, supply the name of a GMT color master dynamic CPT. -C

F : box : – Str – Draws a rectangular border around the scale. -F

G : truncate : – Str – Truncate the incoming CPT so that the lowest and highest z-levels are to zlo and zhi. -G

I : shade : – Number or [] – Add illumination effects. -I

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

Jz : z_axis : – Str – Set z-axis scaling. -Jz

L : equal_size : – Str or [] – Gives equal-sized color rectangles. Default scales rectangles according to the z-range in the CPT. -L

M : monochrome : – Bool or [] – Force conversion to monochrome image using the (television) YIQ transformation. -M

N : dpi : – Str or number – Controls how the color scale is represented by the PostScript language. -N

Q : log : – Str – Selects a logarithmic interpolation scheme [Default is linear]. -Q

R : region : limits : – Str or list – 'xmin/xmax/ymin/ymax[+r][+uunit]'. Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

S : nolines : – Bool or [] – Do not separate different color intervals with black grid lines.

U : stamp : – Str or Bool or [] – '[[just]/dx/dy/][c|label]' Draw GMT time stamp logo on plot.

V : verbose : – Bool or Str – '[level]' Select verbosity level -V

W : zscale : – Number – Multiply all z-values in the CPT by the provided scale. -W

Z : zfile : – Str – File with colorbar-width per color entry. -Z

source

GMT.grdimage — Function.

grdimage(cmd0::String="", arg1=[], arg2=[], arg3=[], arg4=[]; data=[],
		 fmt="", K=false, O=false, first=true, kwargs...)

Produces a gray-shaded (or colored) map by plotting rectangles centered on each grid node and assigning them a gray-shade (or color) based on the z-value.

Full option list at grdimage

Parameters

A : img_out : image_out : – Str – Save an image in a raster format instead of PostScript. -A

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : cmap : – Str – Name of the CPT (for grd_z only). Alternatively, supply the name of a GMT color master dynamic CPT. -C

D : img_in : image_in : – Str or [] – Specifies that the grid supplied is an image file to be read via GDAL. -D

E : dpi : – Int or [] – Sets the resolution of the projected grid that will be created. -E

G : – Str or Int – -G

I : shade : intensity : intensfile : – Str or GMTgrid – Gives the name of a grid file or GMTgrid with intensities in the (-1,+1) range, or a grdgradient shading flags. -I

M : monochrome : – Bool or [] – Force conversion to monochrome image using the (television) YIQ transformation. -M

N : noclip : – Bool or [] – Do not clip the image at the map boundary. -N

Q : nan_t : nan_alphan : – Bool or [] – Make grid nodes with z = NaN transparent, using the colormasking feature in PostScript Level 3.

R : region : limits : – Str or list – 'xmin/xmax/ymin/ymax[+r][+uunit]'. Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

U : stamp : – Str or Bool or [] – '[[just]/dx/dy/][c|label]' Draw GMT time stamp logo on plot.

V : verbose : – Bool or Str – '[level]' Select verbosity level -V

X : x_offset : – Str – '[a|c|f|r][x-shift[u]]'

Y : y_offset : – Str – '[a|c|f|r][y-shift[u]]' Shift plot origin. -Y

f : colinfo : – Str –

n : interp : – Str – -n

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.grdcontour — Function.

grdcontour(cmd0::String="", arg1=[]; data=[], fmt="", K=false, O=false,
           first=true, kwargs...)

Reads a 2-D grid file or a GMTgrid type and produces a contour map by tracing each contour through the grid.

Full option list at pscontour

Parameters

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

A : annot : – Str or Number – Save an image in a raster format instead of PostScript. -A

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : cont : contour : – Str or Number – Contours to be drawn. -C

D : dump : – Str – Dump contours as data line segments; no plotting takes place. -D

F : force : – Str or [] – Force dumped contours to be oriented so that higher z-values are to the left (-Fl [Default]) or right. -F

G : labels : – Str – Controls the placement of labels along the quoted lines. -G

Jz : z_axis : – Str – Set z-axis scaling. -Jz

L : range : – Str – Limit range: Do not draw contours for data values below low or above high. -L

Q : cut : – Str or Number – Do not draw contours with less than cut number of points. -Q

S : smooth : – Number – Used to resample the contour lines at roughly every (gridbox_size/smoothfactor) interval. -S

T : ticks : – Str – Draw tick marks pointing in the downward direction every gap along the innermost closed contours. -T

R : region : limits : – Str or list – 'xmin/xmax/ymin/ymax[+r][+uunit]'. Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

U : stamp : – Str or Bool or [] – '[[just]/dx/dy/][c|label]' Draw GMT time stamp logo on plot.

V : verbose : – Bool or Str – '[level]' Select verbosity level -V

W : pen : – Str or Number – Sets the attributes for the particular line. -W

X : x_offset : – Str – '[a|c|f|r][x-shift[u]]'

Y : y_offset : – Str – '[a|c|f|r][y-shift[u]]' Shift plot origin. -Y

Z : scale : – Str – Use to subtract shift from the data and multiply the results by factor before contouring starts. -Z

bo : binary_out : – Str –

do : nodata_out : – Number – Examine all output columns and if any item equals NAN substitute it with the chosen missing data value -do

e : patern : – Str –

f : colinfo : – Str –

h : headers : – Str –

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.grdview — Function.

grdview(cmd0::String="", arg1=[], arg2=[], arg3=[], arg4=[], arg5=[], arg6=[]; data=[],
        fmt="", K=false, O=false, first=true, kwargs...)

Reads a 2-D grid file and produces a 3-D perspective plot by drawing a mesh, painting a colored/grayshaded surface made up of polygons, or by scanline conversion of these polygons to a raster image.

Full option list at grdview

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – 'xmin/xmax/ymin/ymax[+r][+uunit]'. Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : cmap : – Str – Name of the CPT (for grd_z only). Alternatively, supply the name of a GMT color master dynamic CPT. -C

G : drapefile : – Str or GMTgrid or a Tuple with 3 GMTgrid types – Drape the image in drapefile on top of the relief provided by relief_file. -G

I : shade : intensity : intensfileintens : – Str or GMTgrid – Gives the name of a grid file or GMTgrid with intensities in the (-1,+1) range, or a grdgradient shading flags. -I

Jz : z_axis : – Str – Set z-axis scaling. -Jz

N : plane : – Str or Int – Draws a plane at this z-level. -N

Q : type : – Str or Int – Specify m for mesh plot, s* for surface, **i for image. -Q

S : smooth : – Number – Smooth the contours before plotting. -S

T : no_interp : – Str – Plot image without any interpolation. -T

W : contour : mesh : facade : – Str – Draw contour, mesh or facade. Append pen attributes. -W

U : stamp : – Str or Bool or [] – '[[just]/dx/dy/][c|label]' Draw GMT time stamp logo on plot.

V : verbose : – Bool or Str – '[level]' Select verbosity level -V

X : x_offset : – Str – '[a|c|f|r][x-shift[u]]'

Y : y_offset : – Str – '[a|c|f|r][y-shift[u]]' Shift plot origin. -Y

f : colinfo : – Str –

n : interp : – Str – -n

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.makecpt — Function.

makecpt(cmd0::String="", arg1=[]; data=[], kwargs...)

Make static color palette tables (CPTs).

Full option list at makecpt

A : alpha : transparency : – Str – Sets a constant level of transparency (0-100) for all color slices. -A

C : color : cmap : – Str – Name of the CPT (for grd_z only). Alternatively, supply the name of a GMT color master dynamic CPT. -C

D : – Str or [] – Select the back- and foreground colors to match the colors for lowest and highest z-values in the output CPT. -D

E : data_levels : – Int or [] – Implies reading data table(s) from file or arrays. We use the last data column to determine the data range -E

F : force : – Str – Force output CPT to written with r/g/b codes, gray-scale values or color name. -F

G : truncate : – Str – Truncate the incoming CPT so that the lowest and highest z-levels are to zlo and zhi. -G

I : inverse : reverse : – Str – Reverse the sense of color progression in the master CPT. -I

N : – Bool or [] – Do not write out the background, foreground, and NaN-color fields.

Q : log : – Str – Selects a logarithmic interpolation scheme [Default is linear]. -Q

T : range : – Str – Defines the range of the new CPT by giving the lowest and highest z-value and interval. -T

W : wrap : categorical : – Bool or [] – Do not interpolate the input color table but pick the output colors starting at the beginning of the color table, until colors for all intervals are assigned. -W

Z : continuous : – Bool or [] – Creates a continuous CPT [Default is discontinuous, i.e., constant colors for each interval]. -Z

source