Home

GMT.jl Documentation

Index

Examples

Here's the "Hello World"

And the "Hello Round World"

Color images

Perspective view

GMT.jl Documentation

Index

Functions

Index

GMT.grdcontour

GMT.grdimage

GMT.grdview

GMT.makecpt

GMT.pscoast

GMT.psscale

GMT.psxy

Functions

GMT.psxy — Function.

psxy(cmd0::String="", arg1=[]; caller=[], data=[], fmt="", K=false, O=false,
     first=true, kwargs...)

reads (x,y) pairs from files [or standard input] and generates PostScript code that will plot lines, polygons, or symbols at those locations on a map.

Full option list at psxy

Parameters

A : straight_lines : – Str – By default, geographic line segments are drawn as great circle arcs. To draw them as straight lines, use the -A flag.

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – 'xmin/xmax/ymin/ymax[+r][+uunit]'. Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : – Str – Give a CPT or specify -Ccolor1,color2[,color3,...] to build a linear continuous CPT from those colors automatically. -C

D : offset : – Str – 'dx/dy' Offset the plot symbol or line locations by the given amounts dx/dy.

E : error_bars : – Str – '[x|y|X|Y][+a][+cl|f][+n][+wcap][+ppen]' Draw symmetrical error bars. -E

F : conn : connection : – Str – '[c|n|r][a|f|s|r|refpoint]' Alter the way points are connected -F

G : fill : markerfacecolor : MarkerFaceColor : – Str – Select color or pattern for filling of symbols or polygons. BUT WARN: the alias 'fill' will set the color of polygons OR symbols but not the two together. If your plot has polygons and symbols, use 'fill' for the polygons and 'markerfacecolor' for filling the symbols. Same applyies for W bellow -G

I : intens : – Str or number – Use the supplied intens value (in the [-1 1] range) to modulate the fill color by simulating illumination.

L : closed_polygon : – Str – '[+b|d|D][+xl|r|x0][+yl|r|y0][+ppen]' Force closed polygons. -L

N : no_clip : –- Str or [] – '[c|r]' Do NOT clip symbols that fall outside map border

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

S : symbol : marker : Marker : – Str – '[symbol][size[u]]' Plot symbols (including vectors, pie slices, fronts, decorated or quoted lines). -S Alternatively select a sub-set of symbols using the aliases: marker or Marker and values:

-, x_dash

+, plus

a, *, star

c, circle

d, diamond

g, octagon

h, hexagon

i, v, inverted_tri

n, pentagon

p, ., point

r, rectangle

s, square

t, ^, triangle

x, cross

y, y_dash

W : line_attribs : markeredgecolor : MarkerEdgeColor : – Str – '[pen][attr]' Set pen attributes for lines or the outline of symbols -W WARNING: the pen attributes will set the pen of polygons OR symbols but not the two together. If your plot has polygons and symbols, use W or line_attribs for the polygons and markeredgecolor or MarkerEdgeColor for filling the symbols. Similar to S above.

U : stamp : – Str or Bool or [] – '[[just]/dx/dy/][c|label]' Draw GMT time stamp logo on plot.

V : verbose : – Bool or Str – '[level]' Select verbosity level -V

X : x_offset : – Str – '[a|c|f|r][x-shift[u]]'

Y : y_offset : – Str – '[a|c|f|r][y-shift[u]]' Shift plot origin. -Y

a : aspatial : – Str –

bi : binary_in : – Str –

di : nodata_in : – Str –

e : patern : – Str –

f : colinfo : – Str –

g : gaps : – Str –

h : headers : – Str –

i : input_col : – Str –

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.pscoast — Function.

pscoast(cmd0::String=""; fmt="", clip=[], K=false, O=false, first=true, kwargs...)

Plot continents, shorelines, rivers, and borders on maps. Plots grayshaded, colored, or textured land-masses [or water-masses] on maps and [optionally] draws coastlines, rivers, and political boundaries. A map projection must be supplied.

Full option list at http://gmt.soest.hawaii.edu/doc/latest/pscoast.html

- F = box
- M = dump
- Td = rose
- Tm = compass

Parameters
----------

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – 'xmin/xmax/ymin/ymax[+r][+uunit]'. Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

A : area : – Str or number – 'min_area[/min_level/max_level][+ag|i|s|S][+r|l][+ppercent]' Features with an area smaller than min_area in km^2 or of hierarchical level that is lower than min_level or higher than max_level will not be plotted.

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : river_fill : – Str – Set the shade, color, or pattern for lakes and river-lakes.

D : res : resolution : – Str – Selects the resolution of the data set to use ((f)ull, (h)igh, (i)ntermediate, (l)ow, and (c)rude).

E : ECW : – Str – Tuple(Str, Str); Tuple("code", (pen)), ex: ("PT",(0.5,"red","–")); Tuple((...),(...),...) 'code1,code2,...[+l|L][+gfill][+ppen]' Select painting or dumping country polygons from the Digital Chart of the World

G : land : – Str – Select filling or clipping of “dry” areas.

I : rivers : – Str – 'river[/pen]' Draw rivers. Specify the type of rivers and [optionally] append pen attributes.

L : map_scale : – Str – Dtraw a map scale.

N : borders : – Str – 'border[/pen]' Draw political boundaries. Specify the type of boundary and [optionally] append pen attributes

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

S : water : – Str – Select filling or clipping of “wet” areas.

U : stamp : – Str or Bool or [] – '[[just]/dx/dy/][c|label]' Draw GMT time stamp logo on plot.

V : verbose : – Bool or Str – '[level]' Select verbosity level -V

W : shore : – Str – '[level/]pen' Draw shorelines [Default is no shorelines]. Append pen attributes.

X : x_offset : – Str – '[a|c|f|r][x-shift[u]]'

Y : y_offset : – Str – '[a|c|f|r][y-shift[u]]' Shift plot origin. -Y

bo : binary_out : – Str –

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.psscale — Function.

psscale(cmd0::String="", arg1=[]; fmt="", K=false, O=false, first=true, kwargs...)

source

GMT.grdimage — Function.

grdimage(cmd0::String="", arg1=[], arg2=[], arg3=[], arg4=[]; data=[],
		 fmt="", K=false, O=false, first=true, kwargs...)

Produces a gray-shaded (or colored) map by plotting rectangles centered on each grid node and assigning them a gray-shade (or color) based on the z-value.

Full option list at grdimage

Parameters

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – 'xmin/xmax/ymin/ymax[+r][+uunit]'. Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : cmap : – Str – Name of the CPT (for grd_z only). Alternatively, supply the name of a GMT color master dynamic CPT. -C

U : stamp : – Str or Bool or [] – '[[just]/dx/dy/][c|label]' Draw GMT time stamp logo on plot.

V : verbose : – Bool or Str – '[level]' Select verbosity level -V

X : x_offset : – Str – '[a|c|f|r][x-shift[u]]'

Y : y_offset : – Str – '[a|c|f|r][y-shift[u]]' Shift plot origin. -Y

source

GMT.grdcontour — Function.

grdcontour(cmd0::String="", arg1=[]; data=[], fmt="", K=false, O=false, first=true, kwargs...)

Parameters

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – 'xmin/xmax/ymin/ymax[+r][+uunit]'. Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

U : stamp : – Str or Bool or [] – '[[just]/dx/dy/][c|label]' Draw GMT time stamp logo on plot.

V : verbose : – Bool or Str – '[level]' Select verbosity level -V

X : x_offset : – Str – '[a|c|f|r][x-shift[u]]'

Y : y_offset : – Str – '[a|c|f|r][y-shift[u]]' Shift plot origin. -Y

source

GMT.grdview — Function.

grdview(cmd0::String="", arg1=[], arg2=[], arg3=[], arg4=[], arg5=[], arg6=[]; data=[],
        fmt="", K=false, O=false, first=true, kwargs...)

Reads a 2-D grid file and produces a 3-D perspective plot by drawing a mesh, painting a colored/grayshaded surface made up of polygons, or by scanline conversion of these polygons to a raster image.

Full option list at grdimage

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – 'xmin/xmax/ymin/ymax[+r][+uunit]'. Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : cmap : – Str – Name of the CPT (for grd_z only). Alternatively, supply the name of a GMT color master dynamic CPT. -C

U : stamp : – Str or Bool or [] – '[[just]/dx/dy/][c|label]' Draw GMT time stamp logo on plot.

V : verbose : – Bool or Str – '[level]' Select verbosity level -V

X : x_offset : – Str – '[a|c|f|r][x-shift[u]]'

Y : y_offset : – Str – '[a|c|f|r][y-shift[u]]' Shift plot origin. -Y

n : interp : – Str – -n

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.makecpt — Function.

makecpt(cmd0::String="", arg1=[]; data=[], kwargs...)

Make static color palette tables (CPTs).

julia> cpt = makecpt(range="-1/1/0.1");

julia> (size(cpt.colormap,1) == 20) && (cpt.colormap[1,:] == [0.875, 0.0, 1.0])
true

A : alpha : transparency : – Str –

Sets a constant level of transparency (0-100) for all color slices.
[`-A`](http://gmt.soest.hawaii.edu/doc/latest/makecpt.html#a)

C : color : cmap : – Str – Name of the CPT (for grd_z only). Alternatively, supply the name of a GMT color master dynamic CPT. -C

D : – Str or [] –

Select the back- and foreground colors to match the colors for lowest and highest
z-values in the output CPT. [`-D`](http://gmt.soest.hawaii.edu/doc/latest/makecpt.html#d)

E : data_levels : – Int or [] –

Implies reading data table(s) from file or arrays. We use the last data column to
determine the data range
[`-E`](http://gmt.soest.hawaii.edu/doc/latest/makecpt.html#e)

F : force : – Str – Force output CPT to written with r/g/b codes, gray-scale values or color name.

[`-F`](http://gmt.soest.hawaii.edu/doc/latest/makecpt.html#f)

G : truncate : – Str – Truncate the incoming CPT so that the lowest and highest z-levels are to zlo and zhi.

[`-G`](http://gmt.soest.hawaii.edu/doc/latest/makecpt.html#g)

I : inverse : reverse : – Str – Reverse the sense of color progression in the master CPT.

[`-I`](http://gmt.soest.hawaii.edu/doc/latest/makecpt.html#i)

N : – Bool or [] – Do not write out the background, foreground, and NaN-color fields.

Q : log : – Str – Selects a logarithmic interpolation scheme [Default is linear].

[`-Q`](http://gmt.soest.hawaii.edu/doc/latest/makecpt.html#q)

T : range : – Str – Defines the range of the new CPT by giving the lowest and highest z-value and interval.

[`-T`](http://gmt.soest.hawaii.edu/doc/latest/makecpt.html#t)

W : wrap : categorical : – Bool or [] –

Do not interpolate the input color table but pick the output colors starting at the
beginning of the color table, until colors for all intervals are assigned.
[`-W`](http://gmt.soest.hawaii.edu/doc/latest/makecpt.html#w)

Z : continuous : – Bool or [] – Creates a continuous CPT [Default is discontinuous, i.e., constant colors for each interval].

[`-Z`](http://gmt.soest.hawaii.edu/doc/latest/makecpt.html#z)

source