Home

GMT.jl Documentation

Monolitic

How input and output are assigned

Introduction

Examples

Here's the "Hello World"

And the "Hello Round World"

Simple contours

Color images

Perspective view

Warp image in geographical projection

GMT.jl Documentation

Index

Functions

By Modules

Specifying the pen attributes

Specifying the axes

Specifying the figure size

The output format

The GMT.jl types

Grid type

Image type

Dataset type

CPT type

Postscript type

Index

GMT.coast

GMT.gmtinfo

GMT.grdcontour

GMT.grdimage

GMT.grdinfo

GMT.grdtrack

GMT.grdview

GMT.histogram

GMT.imshow

GMT.makecpt

GMT.nearneighbor

GMT.plot

GMT.rose

GMT.scale

GMT.solar

GMT.surface

GMT.text

GMT.triangulate

GMT.xy

Functions

GMT.plot — Method.

plot(arg1::Array; fmt="", kwargs...)

reads (x,y) pairs from files [or standard input] and generates PostScript code that will plot lines, polygons, or symbols at those locations on a map.

Full option list at psxy

Parameters

A : straight_lines : – Str – By default, geographic line segments are drawn as great circle arcs. To draw them as straight lines, use the -A flag.

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : – Str – Give a CPT or specify -Ccolor1,color2[,color3,...] to build a linear continuous CPT from those colors automatically. -C

D : offset : – Str – Offset the plot symbol or line locations by the given amounts dx/dy.

E : error_bars : – Str – Draw symmetrical error bars. -E

F : conn : connection : – Str – Alter the way points are connected -F

G : fill : markerfacecolor : MarkerFaceColor : – Str – Select color or pattern for filling of symbols or polygons. BUT WARN: the alias 'fill' will set the color of polygons OR symbols but not the two together. If your plot has polygons and symbols, use 'fill' for the polygons and 'markerfacecolor' for filling the symbols. Same applyies for W bellow -G

I : intens : – Str or number – Use the supplied intens value (in the [-1 1] range) to modulate the fill color by simulating illumination.

L : closed_polygon : – Str – Force closed polygons. -L

N : no_clip : –- Str or [] – Do NOT clip symbols that fall outside map border

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

S : symbol : marker : Marker : – Str – Plot symbols (including vectors, pie slices, fronts, decorated or quoted lines). -S Alternatively select a sub-set of symbols using the aliases: marker or Marker and values:

-, x_dash

+, plus

a, *, star

c, circle

d, diamond

g, octagon

h, hexagon

i, v, inverted_tri

n, pentagon

p, ., point

r, rectangle

s, square

t, ^, triangle

x, cross

y, y_dash

W : line_attribs : markeredgecolor : MarkerEdgeColor : – Str – Set pen attributes for lines or the outline of symbols -W WARNING: the pen attributes will set the pen of polygons OR symbols but not the two together. If your plot has polygons and symbols, use W or line_attribs for the polygons and markeredgecolor or MarkerEdgeColor for filling the symbols. Similar to S above.

U : stamp : – Str or Bool or [] – Draw GMT time stamp logo on plot. -U

V : verbose : – Bool or Str – Select verbosity level -V

X : x_offset : – Str –

Y : y_offset : – Str – Shift plot origin. -Y

a : aspatial : – Str – -a

bi : binary_in : – Str – -bi

di : nodata_in : – Str – -di

e : pattern : – Str – -e

f : colinfo : – Str – -f

g : gaps : – Str – -g

h : headers : – Str – -h

i : input_col : – Str – -i

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.imshow — Method.

imshow(arg1; fmt="",  kw...)

Is a simple front end to the grdimage program that accepts GMTgrid, GMTimage, 2D array of floats or strings with file names of grids or images. The normal options of the grdimage program also apply here but some clever guessing of suitable necessary parameters is done if they are not provided. Contrary to other image producing modules the "show' keyword is not necessary to display the image. Here it is set by default. If user wants to use imshow to create layers of a more complex fig he can use show=false for the intermediate layers.

Examples

# Plot vertical shaded illuminated view of the Mexican hat
julia> G = gmt("grdmath -R-15/15/-15/15 -I0.3 X Y HYPOT DUP 2 MUL PI MUL 8 DIV COS EXCH NEG 10 DIV EXP MUL =");
julia> imshow(G, frame="a", shade="+a45")

# Plot a random heat map
julia> imshow(rand(128,128))

# Display a web downloaded jpeg image wrapped into a sinusoidal projection
julia> imshow("http://larryfire.files.wordpress.com/2009/07/untooned_jessicarabbit.jpg", region="d", frame="g", proj="I15", img_in="r", fmt="jpg") 

See also: grdimage

source

GMT.coast — Function.

coast(cmd0::String=""; fmt="", clip=[], kwargs...)

Plot continents, shorelines, rivers, and borders on maps. Plots grayshaded, colored, or textured land-masses [or water-masses] on maps and [optionally] draws coastlines, rivers, and political boundaries. A map projection must be supplied.

Full option list at pscoast

Parameters

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

A : area : – Str or Number – Features with an area smaller than min_area in km^2 or of hierarchical level that is lower than min_level or higher than max_level will not be plotted. -A

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : river_fill : – Str – Set the shade, color, or pattern for lakes and river-lakes. -C

D : res : resolution : – Str – Selects the resolution of the data set to use ((f)ull, (h)igh, (i)ntermediate, (l)ow, and (c)rude). -D

E : DCW : – Str – Select painting or dumping country polygons from the Digital Chart of the World. -E

Tuple("code", Str); Tuple("code" [,"fill"], (pen)); Tuple((...),(...),...)

ex: ("PT",(0.5,"red","–")); (("PT","gblue",(0.5,"red"),("ES",(0.5,"yellow")))

F : box : – Str – Draws a rectangular border around the map scale or rose. -F

G : land : – Str – Select filling or clipping of “dry” areas. -G

I : rivers : – Str – Draw rivers. Specify the type of rivers and [optionally] append pen attributes. -I

L : map_scale : – Str – Draw a map scale. -L

M : dump : – Str – Dumps a single multisegment ASCII output. No plotting occurs. -M

N : borders : – Str – Draw political boundaries. Specify the type of boundary and [optionally] append pen attributes -N

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

S : water : – Str – Select filling or clipping of “wet” areas. -S

Td : rose` : – Str – Draws a map directional rose on the map at the location defined by the reference and anchor points. -Td

Tm : compass : – Str – Draws a map magnetic rose on the map at the location defined by the reference and anchor points. -Tm

U : stamp : – Str or Bool or [] – Draw GMT time stamp logo on plot. -U

V : verbose : – Bool or Str – Select verbosity level -V

W : shore : – Str – Draw shorelines [Default is no shorelines]. Append pen attributes. -W

X : x_offset : – Str –

Y : y_offset : – Str – Shift plot origin. -Y

bo : binary_out : – Str – -bo

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.histogram — Function.

histogram(cmd0::String="", arg1=[]; fmt::String="", kwargs...)

Reads file and examines the first data column to calculate histogram parameters based on the bin-width provided.

Full option list at pshistogram

Parameters

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

W : bin : width : – Number or Str –

Sets the bin width used for histogram calculations.
[`-W`](http://gmt.soest.hawaii.edu/doc/latest/pshistogram.html#w)

A : horizontal : – Bool or [] –

Plot the histogram horizontally from x = 0 [Default is vertically from y = 0].
[`-A`](http://gmt.soest.hawaii.edu/doc/latest/pshistogram.html#a)

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : – Str or GMTcpt –

Give a CPT. The mid x-value for each bar is used to look-up the bar color.
[`-C`](http://gmt.soest.hawaii.edu/doc/latest/pshistogram.html#c)

D : annot : annotate : – Str or [] –

Annotate each bar with the count it represents.
[`-D`](http://gmt.soest.hawaii.edu/doc/latest/pshistogram.html#d)

F : center : – Bool or [] –

Center bin on each value. [Default is left edge].
[`-F`](http://gmt.soest.hawaii.edu/doc/latest/pshistogram.html#f)

G : fill : – Number or Str –

Select filling of bars [if no G, L or C set G=100].
[`-G`](http://gmt.soest.hawaii.edu/doc/latest/pshistogram.html#g)

I : inquire : – Bool or [] –

Inquire about min/max x and y after binning.
[`-I`](http://gmt.soest.hawaii.edu/doc/latest/pshistogram.html#i)

L : labels : – Str or [] –

Draw bar outline using the specified pen thickness [if no G, L or C set L=0.5].
[`-L`](http://gmt.soest.hawaii.edu/doc/latest/pshistogram.html#l)

N : normal : – Str –

Draw the equivalent normal distribution; append desired pen [0.5p,black].
[`-N`](http://gmt.soest.hawaii.edu/doc/latest/pshistogram.html#n)

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

Q : alpha : – Number or [] –

Sets the confidence level used to determine if the mean resultant is significant.
[`-Q`](http://gmt.soest.hawaii.edu/doc/latest/pshistogram.html#q)

R : region : – Str –

Specifies the ‘region’ of interest in (r,azimuth) space. r0 is 0, r1 is max length in units.
[`-R`](http://gmt.soest.hawaii.edu/doc/latest/pshistogram.html#r)

S : stairs : – Str or number –

Draws a stairs-step diagram which does not include the internal bars of the default histogram.
[`-S`](http://gmt.soest.hawaii.edu/doc/latest/pshistogram.html#s)

Z : kind : – Number or Str –

Choose between 6 types of histograms.
[`-Z`](http://gmt.soest.hawaii.edu/doc/latest/pshistogram.html#z)

U : stamp : – Str or Bool or [] – Draw GMT time stamp logo on plot. -U

V : verbose : – Bool or Str – Select verbosity level -V

X : x_offset : – Str –

Y : y_offset : – Str – Shift plot origin. -Y

bi : binary_in : – Str – -bi

di : nodata_in : – Str – -di

e : pattern : – Str – -e

h : headers : – Str – -h

i : input_col : – Str – -i

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

swap_xy : Str or Bool or [] – -:

source

GMT.scale — Function.

scale(cmd0::String="", arg1=[]; fmt="", kwargs...)

Plots gray scales or color scales on maps.

Full option list at psscale

D : position : – Str – Defines the reference point on the map for the color scale using one of four coordinate systems. -D

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : cmap : – Str – Name of the CPT (for grd_z only). Alternatively, supply the name of a GMT color master dynamic CPT. -C

F : box : – Str – Draws a rectangular border around the scale. -F

G : truncate : – Str – Truncate the incoming CPT so that the lowest and highest z-levels are to zlo and zhi. -G

I : shade : – Number or [] – Add illumination effects. -I

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

Jz : z_axis : – Str – Set z-axis scaling. -Jz

L : equal_size : – Str or [] – Gives equal-sized color rectangles. Default scales rectangles according to the z-range in the CPT. -L

M : monochrome : – Bool or [] – Force conversion to monochrome image using the (television) YIQ transformation. -M

N : dpi : – Str or number – Controls how the color scale is represented by the PostScript language. -N

Q : log : – Str – Selects a logarithmic interpolation scheme [Default is linear]. -Q

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

S : nolines : – Bool or [] – Do not separate different color intervals with black grid lines.

U : stamp : – Str or Bool or [] – Draw GMT time stamp logo on plot. -U

V : verbose : – Bool or Str – Select verbosity level -V

W : zscale : – Number – Multiply all z-values in the CPT by the provided scale. -W

Z : zfile : – Str – File with colorbar-width per color entry. -Z

source

GMT.text — Function.

text(cmd0::String="", arg1=[]; fmt="", kwargs...)

Plots text strings of variable size, font type, and orientation. Various map projections are provided, with the option to draw and annotate the map boundaries.

Full option list at pstext

Parameters

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

A : azimuths : – Bool or [] – Angles are given as azimuths; convert them to directions using the current projection. -A

C : clearance : – Str – Sets the clearance between the text and the surrounding box [15%]. -C

D : offset : – Str – Offsets the text from the projected (x,y) point by dx,dy [0/0]. -D

F : text_attrib : – Str or number – Specify up to three text attributes (font, angle, and justification). -F

G : fill : – Number or Str – Sets the shade or color used for filling the text box [Default is no fill]. -G

Jz : z_axis : – Str – Set z-axis scaling. -Jz

L : list : – Bool or [] – Lists the font-numbers and font-names available, then exits. -L

N : no_clip : –- Str or [] – Do NOT clip text at map boundaries. -N

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

Q : change_case : –- Str – Change all text to either lower or upper case. -Q

T : text_box : –- Str – Specify the shape of the textbox when using G and/or W. -T

W : line_attribs : – Str – Sets the pen used to draw a rectangle around the text string. -W

Z : threeD : – Str – For 3-D projections: expect each item to have its own level given in the 3rd column. -Z

U : stamp : – Str or Bool or [] – Draw GMT time stamp logo on plot. -U

V : verbose : – Bool or Str – Select verbosity level -V

X : x_offset : – Str –

Y : y_offset : – Str – Shift plot origin. -Y

a : aspatial : – Str – -a

bi : binary_in : – Str – -bi

di : nodata_in : – Str – -di

e : pattern : – Str – -e

f : colinfo : – Str – -f

g : gaps : – Str – -g

h : headers : – Str – -h

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

swap_xy : Str or Bool or [] – -:

source

GMT.rose — Function.

rose(cmd0::String="", arg1=[]; fmt="", kwargs...)

Reads (length,azimuth) pairs from file and plot a windrose diagram.

Full option list at psrose

Parameters

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

A : sector : – Str or number –

Gives the sector width in degrees for sector and rose diagram.
[`-A`](http://gmt.soest.hawaii.edu/doc/latest/psrose.html#a)

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : vectors : – Str –

Plot vectors showing the principal directions given in the mode_file file.
[`-C`](http://gmt.soest.hawaii.edu/doc/latest/psrose.html#c)

D : shift : – Bool or [] –

Shift sectors so that they are centered on the bin interval (e.g., first sector is centered on 0 degrees).
[`-D`](http://gmt.soest.hawaii.edu/doc/latest/psrose.html#d)

F : no_scale : – Bool or [] –

Do not draw the scale length bar [Default plots scale in lower right corner].
[`-F`](http://gmt.soest.hawaii.edu/doc/latest/psrose.html#f)

G : fill : – Number or Str –

Selects shade, color or pattern for filling the sectors [Default is no fill].
[`-G`](http://gmt.soest.hawaii.edu/doc/latest/psrose.html#g)

I : inquire : – Bool or [] –

Inquire. Computes statistics needed to specify a useful -R. No plot is generated.
[`-I`](http://gmt.soest.hawaii.edu/doc/latest/psrose.html#i)

L : pen : – Number or Str –

Draw bar outline using the specified pen thickness. [Default is no outline].
[`-L`](http://gmt.soest.hawaii.edu/doc/latest/psrose.html#l)

M : – Bool or [] –

Used with -C to modify vector parameters.
[`-M`](http://gmt.soest.hawaii.edu/doc/latest/psrose.html#m)

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

Q : alpha : – Str or [] –

Sets the confidence level used to determine if the mean resultant is significant.
[`-Q`](http://gmt.soest.hawaii.edu/doc/latest/psrose.html#q)

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

S : radius : – Bool or [] –

Specifies radius of plotted circle (append a unit from c|i|p).
[`-S`](http://gmt.soest.hawaii.edu/doc/latest/psrose.html#s)

T : – Bool or [] –

Specifies that the input data are orientation data (i.e., have a 180 degree ambiguity)
instead of true 0-360 degree directions [Default].
[`-T`](http://gmt.soest.hawaii.edu/doc/latest/psrose.html#t)

W : pen : – Str or tuple –

Set pen attributes for sector outline or rose plot. [Default is no outline].
[`-W`](http://gmt.soest.hawaii.edu/doc/latest/psrose.html#w)

Z : scale : – Str –

Multiply the data radii by scale.
[`-Z`](http://gmt.soest.hawaii.edu/doc/latest/psrose.html#z)

U : stamp : – Str or Bool or [] – Draw GMT time stamp logo on plot. -U

V : verbose : – Bool or Str – Select verbosity level -V

X : x_offset : – Str –

Y : y_offset : – Str – Shift plot origin. -Y

bi : binary_in : – Str – -bi

di : nodata_in : – Str – -di

e : pattern : – Str – -e

h : headers : – Str – -h

i : input_col : – Str – -i

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

swap_xy : Str or Bool or [] – -:

source

GMT.solar — Function.

solar(cmd0::String="", arg1=[]; fmt="", kwargs...)

Calculate and plot the day-night terminator and the civil, nautical and astronomical twilights.

Full option list at pssolar

Parameters

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : vectors : – Str – -C

G : fill : – Number or Str – -G

I : inquire : – Bool or [] –

[`-I`](http://gmt.soest.hawaii.edu/doc/latest/pssolar.html#i)

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

M : alpha : – Str or [] –

[`-M`](http://gmt.soest.hawaii.edu/doc/latest/pssolar.html#M)

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

N : radius : – Bool or [] –

[`-N`](http://gmt.soest.hawaii.edu/doc/latest/pssolar.html#n)

T : – Bool or [] –

[`-T`](http://gmt.soest.hawaii.edu/doc/latest/pssolar.html#t)

W : pen : – Str or tuple –

[`-W`](http://gmt.soest.hawaii.edu/doc/latest/pssolar.html#w)

U : stamp : – Str or Bool or [] – Draw GMT time stamp logo on plot. -U

V : verbose : – Bool or Str – Select verbosity level -V

X : x_offset : – Str –

Y : y_offset : – Str – Shift plot origin. -Y

bo : binary_out : – Str – -bo

h : headers : – Str – -h

o : output_col : – Str – -o

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.xy — Function.

xy(cmd0::String="", arg1=[]; fmt="", kwargs...)

reads (x,y) pairs from files [or standard input] and generates PostScript code that will plot lines, polygons, or symbols at those locations on a map.

Full option list at psxy

Parameters

A : straight_lines : – Str – By default, geographic line segments are drawn as great circle arcs. To draw them as straight lines, use the -A flag.

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : – Str – Give a CPT or specify -Ccolor1,color2[,color3,...] to build a linear continuous CPT from those colors automatically. -C

D : offset : – Str – Offset the plot symbol or line locations by the given amounts dx/dy.

E : error_bars : – Str – Draw symmetrical error bars. -E

F : conn : connection : – Str – Alter the way points are connected -F

G : fill : markerfacecolor : MarkerFaceColor : – Str – Select color or pattern for filling of symbols or polygons. BUT WARN: the alias 'fill' will set the color of polygons OR symbols but not the two together. If your plot has polygons and symbols, use 'fill' for the polygons and 'markerfacecolor' for filling the symbols. Same applyies for W bellow -G

I : intens : – Str or number – Use the supplied intens value (in the [-1 1] range) to modulate the fill color by simulating illumination.

L : closed_polygon : – Str – Force closed polygons. -L

N : no_clip : –- Str or [] – Do NOT clip symbols that fall outside map border

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

S : symbol : marker : Marker : – Str – Plot symbols (including vectors, pie slices, fronts, decorated or quoted lines). -S Alternatively select a sub-set of symbols using the aliases: marker or Marker and values:

-, x_dash

+, plus

a, *, star

c, circle

d, diamond

g, octagon

h, hexagon

i, v, inverted_tri

n, pentagon

p, ., point

r, rectangle

s, square

t, ^, triangle

x, cross

y, y_dash

W : line_attribs : markeredgecolor : MarkerEdgeColor : – Str – Set pen attributes for lines or the outline of symbols -W WARNING: the pen attributes will set the pen of polygons OR symbols but not the two together. If your plot has polygons and symbols, use W or line_attribs for the polygons and markeredgecolor or MarkerEdgeColor for filling the symbols. Similar to S above.

U : stamp : – Str or Bool or [] – Draw GMT time stamp logo on plot. -U

V : verbose : – Bool or Str – Select verbosity level -V

X : x_offset : – Str –

Y : y_offset : – Str – Shift plot origin. -Y

a : aspatial : – Str – -a

bi : binary_in : – Str – -bi

di : nodata_in : – Str – -di

e : pattern : – Str – -e

f : colinfo : – Str – -f

g : gaps : – Str – -g

h : headers : – Str – -h

i : input_col : – Str – -i

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

swap_xy : Str or Bool or [] – -:

source

GMT.gmtinfo — Function.

gmtinfo(cmd0::String="", arg1=[]; kwargs...)

Reads its standard input [or from files] and finds the extreme values in each of the columns.

Full option list at gmtinfo

Parameters

A : – Str – Specify how the range should be reported. -A

C : per_column : – Bool or [] – Report the min/max values per column in separate columns [Default uses <min/max> format]. -C

D : center : – Bool or [] – Modifies results obtained by -I by shifting the region to better align with the center of the data. -D

E : get_record : – Str or [] – Returns the record whose column col contains the minimum (l) or maximum (h) value. -E

F : counts : – Str or [] – Returns the counts of various records depending on the appended mode. -F

I : report_region : – Number or Str – Report the min/max of the first n columns to the nearest multiple of the provided increments and output results in the form -Rw/e/s/n -I

L : common_limits : – Bool or [] – Determines common limits across tables or segments. -L

S : for_error_bars : – Str or [] – Add extra space for error bars. Useful together with I option and when later plotting with psxy E. -S

T : nearest_multiple : – Number or Str – Report the min/max of the first (0’th) column to the nearest multiple of dz and output this as the string -Tzmin/zmax/dz. -T

V : verbose : – Bool or Str – Select verbosity level -V

bi : binary_in : – Str – -bi

di : nodata_in : – Str – -di

e : pattern : – Str – -e

f : colinfo : – Str – -f

g : gaps : – Str – -g

h : headers : – Str – -h

i : input_col : – Str – -i

o : output_col : – Str – -o

r : reg : registration : – Str – -r

swap_xy : Str or Bool or [] – -:

source

GMT.grdcontour — Function.

grdcontour(cmd0::String="", arg1=[]; fmt="", kwargs...)

Reads a 2-D grid file or a GMTgrid type and produces a contour map by tracing each contour through the grid.

Full option list at pscontour

Parameters

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

A : annot : – Str or Number – Save an image in a raster format instead of PostScript. -A

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : cont : contours : levels : – Str or Number – Contours to be drawn. -C

D : dump : – Str – Dump contours as data line segments; no plotting takes place. -D

F : force : – Str or [] – Force dumped contours to be oriented so that higher z-values are to the left (-Fl [Default]) or right. -F

G : labels : – Str – Controls the placement of labels along the quoted lines. -G

Jz : z_axis : – Str – Set z-axis scaling. -Jz

L : range : – Str – Limit range: Do not draw contours for data values below low or above high. -L

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

Q : cut : – Str or Number – Do not draw contours with less than cut number of points. -Q

S : smooth : – Number – Used to resample the contour lines at roughly every (gridbox_size/smoothfactor) interval. -S

T : ticks : – Str – Draw tick marks pointing in the downward direction every gap along the innermost closed contours. -T

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

U : stamp : – Str or Bool or [] – Draw GMT time stamp logo on plot. -U

V : verbose : – Bool or Str – Select verbosity level -V

W : pen : – Str or Number – Sets the attributes for the particular line. -W

X : x_offset : – Str –

Y : y_offset : – Str – Shift plot origin. -Y

Z : scale : – Str – Use to subtract shift from the data and multiply the results by factor before contouring starts. -Z

bo : binary_out : – Str – -bo

do : nodata_out : – Number – Examine all output columns and if any item equals NAN substitute it with the chosen missing data value -do

e : pattern : – Str – -e

f : colinfo : – Str – -f

h : headers : – Str – -h

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.grdimage — Function.

grdimage(cmd0::String="", arg1=[], arg2=[], arg3=[]; fmt="", kwargs...)

Produces a gray-shaded (or colored) map by plotting rectangles centered on each grid node and assigning them a gray-shade (or color) based on the z-value.

Full option list at grdimage

Parameters

A : img_out : image_out : – Str – Save an image in a raster format instead of PostScript. -A

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : cmap : – Str – Name of the CPT (for grd_z only). Alternatively, supply the name of a GMT color master dynamic CPT. -C

D : img_in : image_in : – Str or [] – Specifies that the grid supplied is an image file to be read via GDAL. -D

E : dpi : – Int or [] – Sets the resolution of the projected grid that will be created. -E

G : – Str or Int – -G

I : shade : intensity : intensfile : – Str or GMTgrid – Gives the name of a grid file or GMTgrid with intensities in the (-1,+1) range, or a grdgradient shading flags. -I

M : monochrome : – Bool or [] – Force conversion to monochrome image using the (television) YIQ transformation. -M

N : noclip : – Bool or [] – Do not clip the image at the map boundary. -N

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

Q : nan_t : nan_alpha : – Bool or [] – Make grid nodes with z = NaN transparent, using the colormasking feature in PostScript Level 3.

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

U : stamp : – Str or Bool or [] – Draw GMT time stamp logo on plot. -U

V : verbose : – Bool or Str – Select verbosity level -V

X : x_offset : – Str –

Y : y_offset : – Str – Shift plot origin. -Y

f : colinfo : – Str – -f

n : interp : – Str – -n

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.grdinfo — Function.

grdinfo(cmd0::String="", arg1=[]; kwargs...)

Reads a 2-D grid file and reports metadata and various statistics for the (x,y,z) data in the grid file

Full option list at grdinfo

Parameters

C : numeric : – Str or Number – Formats the report using tab-separated fields on a single line. -C

D : tiles : – Number or Str – Divide a single grid’s domain (or the -R domain, if no grid given) into tiles of size dx times dy (set via -I). -D

F : – Bool or [] – Report grid domain and x/y-increments in world mapping format. -F

I : nearest_multiple : – Number or Str – Report the min/max of the region to the nearest multiple of dx and dy, and output this in the form -Rw/e/s/n -I

L : force_scan : – Number or Str – Report stats after actually scanning the data. -L

M : zmin_max : – Bool or [] – Find and report the location of min/max z-values. -M

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

T : nan_t : – Number or Str – Determine min and max z-value. -T

V : verbose : – Bool or Str – Select verbosity level -V

f : colinfo : – Str – -f

source

GMT.grdtrack — Function.

grdtrack(cmd0::String="", arg1=[], arg2=[]; kwargs...)

Interpolates the grid(s) at the positions in the table and writes out the table with the interpolated values added as (one or more) new columns.

Full option list at grdtrack

Parameters

A : interp_path : – Str –

[`-A`](http://gmt.soest.hawaii.edu/doc/latest/grdtrack.html#a)

C : equi : – Str –

[`-C`](http://gmt.soest.hawaii.edu/doc/latest/grdtrack.html#c)

D : dfile : – Str –

[`-D`](http://gmt.soest.hawaii.edu/doc/latest/grdtrack.html#d)

E : by_coord : – Str –

[`-E`](http://gmt.soest.hawaii.edu/doc/latest/grdtrack.html#e)

G : grid : – Str or GMTgrid –

[`-G`](http://gmt.soest.hawaii.edu/doc/latest/grdtrack.html#g)

N : no_skip : – Bool or [] –

[`-N`](http://gmt.soest.hawaii.edu/doc/latest/grdtrack.html#n)

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

S : stack : – Str –

[`-S`](http://gmt.soest.hawaii.edu/doc/latest/grdtrack.html#s)

T : radius : – Number, Str or [] –

[`-T`](http://gmt.soest.hawaii.edu/doc/latest/grdtrack.html#t)

Z : z_only : – Bool or [] –

[`-Z`](http://gmt.soest.hawaii.edu/doc/latest/grdtrack.html#z)

V : verbose : – Bool or Str – Select verbosity level -V

bi : binary_in : – Str – -bi

bo : binary_out : – Str – -bo

di : nodata_in : – Str – -di

e : pattern : – Str – -e

f : colinfo : – Str – -f

g : gaps : – Str – -g

h : headers : – Str – -h

i : input_col : – Str – -i

n : interp : – Str – -n

o : output_col : – Str – -o

s : skip_col : – Str – -s

swap_xy : Str or Bool or [] – -:

'arg1' may contain the table data as an array or a GMTdataset type and GMTgrid or it can be left empty. Same thing for 'arg2'. 'cmd' can have the file name of either the table or the grid.

source

GMT.grdview — Function.

grdview(cmd0::String="", arg1=[], arg2=[], arg3=[]; fmt="", kwargs...)

Reads a 2-D grid file and produces a 3-D perspective plot by drawing a mesh, painting a colored/grayshaded surface made up of polygons, or by scanline conversion of these polygons to a raster image.

Full option list at grdview

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : cmap : – Str – Name of the CPT (for grd_z only). Alternatively, supply the name of a GMT color master dynamic CPT. -C

G : drapefile : – Str or GMTgrid or a Tuple with 3 GMTgrid types – Drape the image in drapefile on top of the relief provided by relief_file. -G

I : shade : intensity : intensfileintens : – Str or GMTgrid – Gives the name of a grid file or GMTgrid with intensities in the (-1,+1) range, or a grdgradient shading flags. -I

Jz : z_axis : – Str – Set z-axis scaling. -Jz

N : plane : – Str or Int – Draws a plane at this z-level. -N

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

Q : type : – Str or Int – Specify m for mesh plot, s* for surface, **i for image. -Q

S : smooth : – Number – Smooth the contours before plotting. -S

T : no_interp : – Str – Plot image without any interpolation. -T

W : contour : mesh : facade : – Str – Draw contour, mesh or facade. Append pen attributes. -W

U : stamp : – Str or Bool or [] – Draw GMT time stamp logo on plot. -U

V : verbose : – Bool or Str – Select verbosity level -V

X : x_offset : – Str –

Y : y_offset : – Str – Shift plot origin. -Y

f : colinfo : – Str – -f

n : interp : – Str – -n

p : view : perspective : – Str – -p

t : alpha : transparency : – Str – -t

source

GMT.makecpt — Function.

makecpt(cmd0::String="", arg1=[]; kwargs...)

Make static color palette tables (CPTs).

Full option list at makecpt

A : alpha : transparency : – Str – Sets a constant level of transparency (0-100) for all color slices. -A

C : color : cmap : – Str – Name of the CPT (for grd_z only). Alternatively, supply the name of a GMT color master dynamic CPT. -C

D : – Str or [] – Select the back- and foreground colors to match the colors for lowest and highest z-values in the output CPT. -D

E : data_levels : – Int or [] – Implies reading data table(s) from file or arrays. We use the last data column to determine the data range -E

F : force : – Str – Force output CPT to written with r/g/b codes, gray-scale values or color name. -F

G : truncate : – Str – Truncate the incoming CPT so that the lowest and highest z-levels are to zlo and zhi. -G

I : inverse : reverse : – Str – Reverse the sense of color progression in the master CPT. -I

N : – Bool or [] – Do not write out the background, foreground, and NaN-color fields.

Q : log : – Str – Selects a logarithmic interpolation scheme [Default is linear]. -Q

T : range : – Str – Defines the range of the new CPT by giving the lowest and highest z-value and interval. -T

W : wrap : categorical : – Bool or [] – Do not interpolate the input color table but pick the output colors starting at the beginning of the color table, until colors for all intervals are assigned. -W

Z : continuous : – Bool or [] – Creates a continuous CPT [Default is discontinuous, i.e., constant colors for each interval]. -Z

bi : binary_in : – Str – -bi

di : nodata_in : – Str – -di

i : input_col : – Str – -i

source

GMT.nearneighbor — Function.

nearneighbor(cmd0::String="", arg1=[]; kwargs...)

Reads arbitrarily located (x,y,z[,w]) triples [quadruplets] and uses a nearest neighbor algorithm to assign an average value to each node that have one or more points within a radius centered on the node. The average value is computed as a weighted mean of the nearest point from each sector inside the search radius. The weighting function used is w(r) = 1 / (1 + d ^ 2), where d = 3 * r / search_radius and r is distance from the node. This weight is modulated by the weights of the observation points [if supplied].

Full option list at nearneighbor

Parameters

I : inc : – Str or Number –

*x_inc* [and optionally *y_inc*] is the grid spacing.
[`-I`](http://gmt.soest.hawaii.edu/doc/latest/nearneighbor.html#i)

N : sectors : – Number or Str – The circular area centered on each node is divided into sectors sectors. -N

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

S : search_radius : – Number – Sets the search_radius that determines which data points are considered close to a node. -S

E : empty : – Bool or [] – Set the value assigned to empty nodes when G is set [NaN]. -E

G : grid : – Str or [] –

Optional output grid file name. If not provided return a GMTgrid type.
[`-G`](http://gmt.soest.hawaii.edu/doc/latest/nearneighbor.html#g)

V : verbose : – Bool or Str – Select verbosity level -V

W : weights : – Bool or [] – Input data have a 4th column containing observation point weights. -W

bi : binary_in : – Str – -bi

di : nodata_in : – Str – -di

e : pattern : – Str – -e

f : colinfo : – Str – -f

h : headers : – Str – -h

i : input_col : – Str – -i

n : interp : – Str – -n

r : reg : registration : – Str – -r

swap_xy : Str or Bool or [] – -:

source

GMT.surface — Function.

surface(cmd0::String="", arg1=[]; kwargs...)

Reads randomly-spaced (x,y,z) triples and produces a binary grid file of gridded values z(x,y) by solving:

(1 - T) * L (L (z)) + T * L (z) = 0

Full option list at surface

Parameters

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

I : inc : – Str or Number –

*x_inc* [and optionally *y_inc*] is the grid spacing.
[`-I`](http://gmt.soest.hawaii.edu/doc/latest/surface.html#i)

A : aspect_ratio : – Number – Aspect ratio. If desired, grid anisotropy can be added to the equations. -A

C : convergence : – Number –

Convergence limit. Iteration is assumed to have converged when the maximum absolute change in any
grid value is less than convergence_limit.
[`-C`](http://gmt.soest.hawaii.edu/doc/latest/surface.html#c)

G : grid : – Str or [] –

Optional output grid file name. If not provided return a GMTgrid type.
[`-G`](http://gmt.soest.hawaii.edu/doc/latest/surface.html#g)

Ll : lower : – Str or Number –

Impose limits on the output solution. lower sets the lower bound. lower can be the name of a grid
file with lower bound values, a fixed value, d to set to minimum input value,
[`-L`](http://gmt.soest.hawaii.edu/doc/latest/surface.html#l)

Lu : upper : – Str or Number – -L

N : max_iterations : – Number –

Number of iterations. Iteration will cease when convergence_limit is reached or when number of
iterations reaches max_iterations.
[`-N`](http://gmt.soest.hawaii.edu/doc/latest/surface.html#n)

Q : suggest : – Bool or [] – Suggest grid dimensions which have a highly composite greatest common factor. -Q

S : search_radius : – Number or Str – Sets the resolution of the projected grid that will be created. -E

T : tension : – Number or Str – Tension factor[s]. These must be between 0 and 1. -T

V : verbose : – Bool or Str – Select verbosity level -V

Z : over_relaxation : – Str or GMTgrid – Over-relaxation factor. This parameter is used to accelerate the convergence; it is a number between 1 and 2. -Z

a : aspatial : – Str – -a

bi : binary_in : – Str – -bi

di : nodata_in : – Str – -di

e : pattern : – Str – -e

f : colinfo : – Str – -f

h : headers : – Str – -h

i : input_col : – Str – -i

r : reg : registration : – Str – -r

swap_xy : Str or Bool or [] – -:

source

GMT.triangulate — Function.

triangulate(cmd0::String="", arg1=[]; kwargs...)

Reads one or more ASCII [or binary] files (or standard input) containing x,y[,z] and performs Delaunay triangulation, i.e., it find how the points should be connected to give the most equilateral triangulation possible.

Full option list at triangulate

Parameters

C : slope_grid : – Number – Read a slope grid (in degrees) and compute the propagated uncertainty in the bathymetry using the CURVE algorithm -C

D : derivatives : – Str – Take either the x- or y-derivatives of surface represented by the planar facets (only used when G is set). -D

E : empty : – Bool or [] – Set the value assigned to empty nodes when G is set [NaN]. -E

G : grid : – Str or [] – Use triangulation to grid the data onto an even grid (specified with R I). Append the name of the output grid file. -G

I : inc : – Str or Number –

*x_inc* [and optionally *y_inc*] is the grid spacing.
[`-I`](http://gmt.soest.hawaii.edu/doc/latest/triangulate.html#i)

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

M : network : – Bool or [] – Output triangulation network as multiple line segments separated by a segment header record. -M

N : ids : – Bool or [] – Used in conjunction with G to also write the triplets of the ids of all the Delaunay vertices -N

Q : voronoi : – Str or [] – Output the edges of the Voronoi cells instead [Default is Delaunay triangle edges] -Q

R : region : limits : – Str or list – Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

S : triangles : – Bool or [] – Output triangles as polygon segments separated by a segment header record. Requires Delaunay triangulation. -S

T : edges : – Bool or [] – Output edges or polygons even if gridding has been selected with the G option -T

V : verbose : – Bool or Str – Select verbosity level -V

Z : xyz : triplets : – Bool or [] – -Z

bi : binary_in : – Str – -bi

bo : binary_out : – Str – -bo

di : nodata_in : – Str – -di

e : pattern : – Str – -e

f : colinfo : – Str – -f

h : headers : – Str – -h

i : input_col : – Str – -i

r : reg : registration : – Str – -r

swap_xy : Str or Bool or [] – -:

source