Home

GMT.jl Documentation

GMT.jl Documentation

Index

Functions

Index

GMT.grdcontour

GMT.grdimage

GMT.grdview

GMT.makecpt

GMT.pscoast

GMT.psscale

GMT.psxy

Functions

GMT.psxy — Function.

psxy(cmd0::String="", arg1=[]; caller=[], data=[], fmt="", K=false, O=false, first=true, kwargs...)

reads (x,y) pairs from files [or standard input] and generates PostScript code that will plot lines, polygons, or symbols at those locations on a map. If a symbol is selected and no symbol size given, then psxy will interpret the third column of the input data as symbol size. Symbols whose size is <= 0 are skipped. If no symbols are specified then the symbol code must be present as last column in the input. If -S is not used, a line connecting the data points will be drawn instead. To explicitly close polygons, use -L. Select a fill with -G. If -G is set, -W will control whether the polygon outline is drawn or not. If a symbol is selected, -G and -W determines the fill and outline/no outline, respectively.

Full option list at psxy

Parameters

A : straight_lines : – Str – By default, geographic line segments are drawn as great circle arcs. To draw them as straight lines, use the -A flag.

J : proj : projection : – Str – Select map projection. Defaults to 12x8 cm with linear (non-projected) maps. -J

R : region : limits : – Str or list – 'xmin/xmax/ymin/ymax[+r][+uunit]'. Specify the region of interest. Set to data minimum BoundinBox if not provided. -R

B : frame : axes : – Str – '[p|s]parameters' Set map boundary frame and axes attributes. -B

C : color : – Str – Give a CPT or specify -Ccolor1,color2[,color3,...] to build a linear continuous CPT from those colors automatically. -C

D : offset : – Str – 'dx/dy' Offset the plot symbol or line locations by the given amounts dx/dy.

E : error_bars : – Str – '[x|y|X|Y][+a][+cl|f][+n][+wcap][+ppen]' Draw symmetrical error bars. -E

F : conn : connection : – Str – '[c|n|r][a|f|s|r|refpoint]' Alter the way points are connected -F

G : fill : markerfacecolor : MarkerFaceColor : – Str – Select color or pattern for filling of symbols or polygons. BUT WARN: the alias 'fill' will set the color of polygons OR symbols but not the two together. If your plot has polygons and symbols, use 'fill' for the polygons and 'markerfacecolor' for filling the symbols. Same applyies for W bellow -G

I : intens : – Str or number – Use the supplied intens value (in the [-1 1] range) to modulate the fill color by simulating illumination.

L : closed_polygon : – Str – '[+b|d|D][+xl|r|x0][+yl|r|y0][+ppen]' Force closed polygons. -L

N : no_clip : –- Str or [] – '[c|r]' Do NOT clip symbols that fall outside map border

P : portrait : –- Bool or [] – Tell GMT to NOT draw in portriat mode (that is, make a Landscape plot)

S : symbol : marker : Marker : – Str – '[symbol][size[u]]' Plot symbols (including vectors, pie slices, fronts, decorated or quoted lines). -S Alternatively select a sub-set of symbols using the aliases: 'marker' or 'Marker' and values:

- or x_dash

+ or plus

a or * or star

c or circle

d or diamond

g or octagon

h or hexagon

i or v or inverted_tri

n or pentagon

p or . or point

r or rectangle

s or square

t or ^ or triangle

x or cross

y or y_dash

W : line_attribs : markeredgecolor : MarkerEdgeColor : – Str – '[pen][attr]' Set pen attributes for lines or the outline of symbols -W WARNING: the pen attributes will set the pen of polygons OR symbols but not the two together. If your plot has polygons and symbols, use 'W' or 'line_attribs' for the polygons and 'markeredgecolor' or 'MarkerEdgeColor' for filling the symbols. Similar to S above.

U : stamp : –- Str or Bool or [] – '[[just]/dx/dy/][c|label]' Draw GMT time stamp logo on plot.

V : verbose : – Bool or Str – '[level]' Select verbosity level -V

X : x_offset : – Str – '[a|c|f|r][x-shift[u]]'

Y : x_offset : – Str – '[a|c|f|r][y-shift[u]]' Shift plot origin. -Y

a : aspatial : – Str –

bi : binary_in : – Str –

di : nodata_in : – Str –

e : patern : – Str –

f : colinfo : – Str –

g : gaps : – Str –

h : headers : – Str –

i : input_col : – Str –

p : perspective : – Str –

t : transparency : – Str –

source

GMT.pscoast — Function.

pscoast(cmd0::String=""; fmt="", clip=[], K=false, O=false, first=true, kwargs...)

Plot continents, shorelines, rivers, and borders on maps. Plots grayshaded, colored, or textured land-masses [or water-masses] on maps and [optionally] draws coastlines, rivers, and political boundaries. A map projection must be supplied.

Full option list at http://gmt.soest.hawaii.edu/doc/latest/pscoast.html

- F = box
- M = dump
- P = portrait
- Td = rose
- Tm = compass
- bo = binary_out
- p = perspective
- t = transparency

Parameters
----------
J : proj : projection : -- Str --
    Select map projection. Defaults to 12x8 cm with linear (non-projected) maps.
    http://gmt.soest.hawaii.edu/doc/latest/psxy.html#j
R : region : limits : -- Str or list --    'xmin/xmax/ymin/ymax[+r][+uunit]'.
    Specify the region of interest. Set to data minimum BoundinBox if not provided.
    http://gmt.soest.hawaii.edu/doc/latest/psxy.html#r
A : area : -- Str or number --
    'min_area[/min_level/max_level][+ag|i|s|S][+r|l][+ppercent]'
    Features with an area smaller than min_area in km^2 or of
    hierarchical level that is lower than min_level or higher than
    max_level will not be plotted.
B : frame : axes : -- Str --  '[p|s]parameters'
    Set map boundary frame and axes attributes.
    http://gmt.soest.hawaii.edu/doc/latest/pscoast.html#b
C : river_fill : -- Str --
    Set the shade, color, or pattern for lakes and river-lakes.
D : res : resolution : -- Str --
    Selects the resolution of the data set to use ((f)ull, (h)igh,
    (i)ntermediate, (l)ow, and (c)rude).
E : ECW : -- Str --  Tuple(Str, Str); Tuple("code", (pen)), ex: ("PT",(0.5,"red","--")); Tuple((...),(...),...)
    'code1,code2,...[+l|L][+gfill][+ppen]'		
    Select painting or dumping country polygons from the Digital Chart of the World
G : land : -- Str --
    Select filling or clipping of “dry” areas.
I : rivers : -- Str --
    'river[/pen]'
    Draw rivers. Specify the type of rivers and [optionally] append pen
    attributes.
L : map_scale : -- Str --
    Dtraw a map scale.
N : borders : -- Str --
    'border[/pen]'
    Draw political boundaries. Specify the type of boundary and
    [optionally] append pen attributes
S : water : -- Str --
    Select filling or clipping of “wet” areas.
U : Str or Bool or []
    Draw GMT time stamp logo on plot.
V : Bool or Str   '[level]'
    Select verbosity level 
	http://gmt.soest.hawaii.edu/doc/latest/psxy.html#v
W : shore : -- Str --
    '[level/]pen'
    Draw shorelines [Default is no shorelines]. Append pen attributes.
X : Str    '[a|c|f|r][x-shift[u]]'
Y : Str    '[a|c|f|r][y-shift[u]]'
    Shift plot origin. 
	http://gmt.soest.hawaii.edu/doc/latest/psxy.html#x

source

GMT.psscale — Function.

psscale(cmd0::String="", arg1=[]; fmt="", K=false, O=false, first=true, kwargs...)

source

GMT.grdimage — Function.

grdimage(cmd0::String="", arg1=[], arg2=[], arg3=[], arg4=[]; data=[],
		 fmt="", K=false, O=false, first=true, kwargs...)

Produces a gray-shaded (or colored) map by plotting rectangles centered on each grid node and assigning them a gray-shade (or color) based on the z-value.

Full option list at http://gmt.soest.hawaii.edu/doc/latest/grdimage.html

Parameters
----------
J : Str
	Select map projection.
R : Str or list
	'xmin/xmax/ymin/ymax[+r][+uunit]'.
	Specify the region of interest.
A : Str or number
	'min_area[/min_level/max_level][+ag|i|s|S][+r|l][+ppercent]'
	Features with an area smaller than min_area in km^2 or of
	hierarchical level that is lower than min_level or higher than
	max_level will not be plotted.
B : Str
	Set map boundary frame and axes attributes.
C : Str
	Set the shade, color, or pattern for lakes and river-lakes.
D : Str
	Selects the resolution of the data set to use ((f)ull, (h)igh,
	(i)ntermediate, (l)ow, and (c)rude).
E : Str; Tuple(Str, Str); Tuple("code", (pen)), ex: ("PT",(0.5,"red","--")); Tuple((...),(...),...)
	'code1,code2,...[+l|L][+gfill][+ppen]'		
	Select painting or dumping country polygons from the Digital Chart of the World
G : Str
	Select filling or clipping of “dry” areas.
I : Str
	'river[/pen]'
	Draw rivers. Specify the type of rivers and [optionally] append pen
	attributes.
N : Str
	'border[/pen]'
	Draw political boundaries. Specify the type of boundary and
	[optionally] append pen attributes
S : Str
	Select filling or clipping of “wet” areas.
U : Str or []
	Draw GMT time stamp logo on plot.
V : Bool or Str   '[level]'
    Select verbosity level 
	http://gmt.soest.hawaii.edu/doc/latest/grdimage.html#v
W : Str
	'[level/]pen'
	Draw shorelines [Default is no shorelines]. Append pen attributes.
X : Str    '[a|c|f|r][x-shift[u]]'
Y : Str    '[a|c|f|r][y-shift[u]]'
    Shift plot origin. 
	http://gmt.soest.hawaii.edu/doc/latest/grdimage.html#x

source

GMT.grdcontour — Function.

grdcontour(cmd0::String="", arg1=[]; data=[], fmt="", K=false, O=false, first=true, kwargs...)

source

GMT.grdview — Function.

grdview(cmd0::String="", arg1=[], arg2=[], arg3=[], arg4=[], arg5=[], arg6=[]; data=[],
        fmt="", K=false, O=false, first=true, kwargs...)

source

GMT.makecpt — Function.

makecpt(cmd0::String="", arg1=[]; data=[], kwargs...)

Make static color palette tables (CPTs).

#jldoctest #julia> cpt = makecpt(range="-1/1/0.1"); #julia> (size(cpt.colormap,1) == 20) && (cpt.colormap[1,:] == [0.875, 0.0, 1.0]) #true #

source